refactor: move translation function and constants in specific files
