+++ src/constants.js
@@ -1,0 +1,25 @@
+// Define available languages
+const LANGUAGES = {
+    en: "English",
+    es: "Spanish",
+    fr: "French",
+    de: "German",
+    it: "Italian",
+    pt: "Portuguese",
+    ru: "Russian",
+    ja: "Japanese",
+    ko: "Korean",
+    zh: "Chinese",
+    ar: "Arabic",
+    hi: "Hindi",
+    nl: "Dutch",
+    pl: "Polish",
+    tr: "Turkish",
+    vi: "Vietnamese",
+    th: "Thai",
+    sv: "Swedish",
+    da: "Danish",
+    fi: "Finnish",
+};
+
+export {LANGUAGES}
